Organize dal files
